+++ activerecord/lib/active_record/relation/batches/batch_enumerator.rb
@@ -5,12 +5,13 @@ module ActiveRecord
     class BatchEnumerator
       include Enumerable
 
-      def initialize(of: 1000, start: nil, finish: nil, relation:, order: :asc, use_ranges: nil) # :nodoc:
-        @of       = of
+      def initialize(of: 1000, start: nil, finish: nil, relation:, cursor_column:, order: :asc, use_ranges: nil) # :nodoc:
+        @of = of
         @relation = relation
         @start = start
         @finish = finish
         @order = order
+        @cursor_column = cursor_column
         @use_ranges = use_ranges
       end
 
@@ -52,7 +53,7 @@ module ActiveRecord
       def each_record(&block)
         return to_enum(:each_record) unless block_given?
 
-        @relation.to_enum(:in_batches, of: @of, start: @start, finish: @finish, load: true, order: @order).each do |relation|
+        @relation.to_enum(:in_batches, of: @of, start: @start, finish: @finish, load: true, cursor_column: @cursor_column, order: @order).each do |relation|
           relation.records.each(&block)
         end
       end
@@ -92,7 +93,7 @@ module ActiveRecord
       #     relation.update_all(awesome: true)
       #   end
       def each(&block)
-        enum = @relation.to_enum(:in_batches, of: @of, start: @start, finish: @finish, load: false, order: @order, use_ranges: @use_ranges)
+        enum = @relation.to_enum(:in_batches, of: @of, start: @start, finish: @finish, load: false, cursor_column: @cursor_column, order: @order, use_ranges: @use_ranges)
         return enum.each(&block) if block_given?
         enum
       end